+++ src/tngsdk/package/helper.py
@@ -30,6 +30,12 @@
 # acknowledge the contributions of their colleagues of the SONATA
 # partner consortium (www.5gtango.eu).
 import copy
+import hashlib
+import os
+import zipfile
+import tempfile
+import time
+import glob
 from tngsdk.package.logger import TangoLogger
 
 
@@ -62,3 +68,74 @@ def dictionary_deep_merge(d1, d2, skip=None):
                 dictionary_deep_merge(d1[k], v)
         else:
             d1[k] = copy.copy(v)
+
+
+def file_hash(path, h_func=hashlib.sha256):
+    h = h_func()
+    with open(path, 'rb', buffering=0) as f:
+        for b in iter(lambda: f.read(128 * 1024), b''):
+            h.update(b)
+    return h.hexdigest()
+
+
+def _makedirs(p):
+    if not os.path.exists(p):
+        LOG.debug("Creating: {}".format(p))
+        os.makedirs(p)
+
+
+def search_for_file(path="**/TOSCA.meta", recursive=True):
+    """
+    Recursively searches for a file.
+    If there are multiple matches, the first one is
+    returned.
+    param: path: src/**/*.c
+    """
+    f_lst = list()
+    try:
+        # Trick17: Depending on the Python version, the recursive
+        # argument might not be available. This provides a fallback.
+        # This reduces the robustness against malformed packages but ensures
+        # that the tool works well in older environments.
+        f_lst = list(glob.iglob(path, recursive=recursive))
+    except BaseException:
+        f_lst = list(glob.iglob(path))
+    LOG.debug("Searching for '{}' found: {}".format(path, f_lst))
+    if len(f_lst) > 0:
+        return f_lst[0]
+    return None
+
+
+def extract_zip_file_to_temp(path_zip, path_dest=None):
+    """
+    Simply extract the entire ZIP file for now.
+    Might be improved for larger files.
+    Always extract to a temp folder to work on.
+    """
+    if path_dest is None:
+        path_dest = tempfile.mkdtemp()
+    LOG.debug("Unzipping '{}' ...".format(path_zip))
+    t_start = time.time()
+    # unzipping
+    with zipfile.ZipFile(path_zip, "r") as f:
+        f.extractall(path_dest)
+    LOG.debug("Unzipping done ({:.4f}s)".format(time.time()-t_start))
+    # get path of output folder (it is based on zip name)
+    wd = find_root_folder_of_pkg(path_dest)
+    LOG.debug("Working root '{}'".format(wd))
+    return wd
+
+
+def find_root_folder_of_pkg(d):
+    """
+    Zip files can be odd and may contain folders
+    w. name of zip file which then contain the root of
+    the package structure.
+    This functions tries to find the right root.
+    """
+    root_indicators = ["TOSCA-Metadata"]
+    for ri in root_indicators:
+        lst = glob.glob("{}/{}".format(d, ri))
+        if len(lst) > 0:
+            return lst[0].replace(ri, "")
+    return d
