+++ src/tngsdk/package/packager/exeptions.py
@@ -1,0 +1,34 @@
+class UnsupportedPackageFormatException(BaseException):
+    pass
+
+
+class MissingInputException(BaseException):
+    pass
+
+
+class MissingMetadataException(BaseException):
+    pass
+
+
+class MissingFileException(BaseException):
+    pass
+
+
+class NapdNotValidException(BaseException):
+    pass
+
+
+class MetadataValidationException(BaseException):
+    pass
+
+
+class ChecksumException(BaseException):
+    pass
+
+
+class InvalidVersionFormat(Exception):
+    pass
+
+
+class NoOSMFilesFound(BaseException):
+    pass
